5222: Request metadata from verify
The bridge should request Metadata from verify on startup, and then
again every few minutes (to ensure it stays fresh). It will use this for
validating the Signatures on AuthnRequests.

Tests use webmock to stub out the endpoint that serves metadata and to
check that the right requests are made. Webmock is started on a dynamic
port to avoid conflicts with other processes.

We considered using @ClassRules for both starting webmock and starting
the dropwizard application in the BridgeApplicationIntegrationTest. But
this would be confusing, because in the test the code we're testing is
the application startup itself (which should make a request for
metadata). To make it clearer what's going on we've switched to using
manual @Before and @After methods to start and stop webmock and we've
manually started the dropwizard application in the test method itself
(since application startup is the code under test).

Authors: @chrisholmes @richardtowers @niyioyelade