5222: Add endpoint to receive AuthnRequests
Most of the handling of AuthnRequests has been pulled into a helper
class, which is lightly covered by unit tests. There are AppRule tests
that post AuthnRequests with valid and invalid signatures to check that
it is being validated correctly.

Since many tests are likely to want Metadata without necessarily caring
about the fact it's requested or the exact payload we've created a
MetadataRule that sets up wiremock with sensible defaults for Metadata.

Need to ensure Authn Request is being signed with the right Cert.
Need to write Code to validate signature on the Authn Request.
Need to ensure the call to metadata on startup is working okay

Authors: @niyioyelade @richardtowers