5222: Wire up bridge to work with compliance tool
Adds a script to initialize the Verify Compliance tool (hosted on our
reference environment) to point at a locally running instance of the
bridge. Changes the bridge to start on a port that doesn't conflict with
Tomcat (or any of the Verify microservices).

Also adds validation of the AuthnRequest Issuer (which has to be hub).

Also removes a dependency on hub-saml-test-utils (which was basically
just making a simple constructor call).

Authors: @richardtowers @suneetkamath