5223: Generate authnrequest to send to connector node
Copied most of the attributes / elements from an example AuthnRequest
generated by the eIDAS stub SP.

Eidas add some SAML extensions which need to be marshalled to and from
XML. There's a RequestedAttribute element which is similar to a
Attribute from opensaml core, but adds an isRequired attribute. This
commit adds the implementation of the opensaml classes needed to
represent, marshal, and unmarshal these elements.

Authors: @richardtowers @sfkamath