5223: Add singleSignOnLocation and submit button to SAML form
This will allow us to POST the SAML to the eidas connector node.
SingleSignOnServiceLocation is pulled from the eidas metadata.

Also did some refatoring of the MetadataRule (you now need one for
verify and one for eidas) and the VerifyEidasBridgeFactory (which is now
more careful about returning the same instances each time).

Authors: @richardtowers @sfkamath @chrisholmes