5223: misc fixes to get the bridge working with eIDAS
It transpires that there are multiple URLs for metadata in eIDAS. The
correct one to use from the point of view of a "service provider" (which
is how the bridge is behaving in this case) is
/ConnectorResponderMetadata. This describes the connector as an IDP and
includes the correct single sign on location for AuthnRequests.
Special thanks to @chrisholmes for noticing this.

Added a default value for signingKeyStore which consists of keys from
IDA Dev PKI. Can't use Strict for the EnvironmentVariableSubstitutor as
the default values from the configuration would not be used

Added a hostname from the config for the bridge metadata gemerator tp
setup an AssertionConsumerService element in the metadata

Added a few extra attributes to the AuthnRequest (as required)

Fixed the form method to be post as it defaulted to get

Authors: @richardtowers @sfkamath @chrisholmes