5405: store IDs in session
When we receive a valid inbound AuthnRequest we need to track the ID inside
it and the accompanying relayState value so that when we generate a Response we
can include these values.  In addition, when generating an AuthnRequest to eIDAS
we need to track the outbound ID so that when we receive a response we can
validate that is correctly in response to our request.

To do this we have adopted a JWT cookie library that we can use to carry signed
versions of the following claims:
* the inbound AuthnRequest ID
* the inbound RelayState
* the outbound AuthnRequest ID