5225: Build response for verify
Validate the response from eIDAS and decrypt / validate the
EncryptedAssertions it contains.  Unpack the Assertion into a plain old
java object representing the bits we need to send to verify.

Transform the details of the assertion into a MatchingDatasetAssertion
for verify (this is a work in progress - just got FirstName at the
moment).

Build a SAML Response for verify and render it in a self-submitting SAML
form.

Add exception mappers to respond with 400 if the request is incorrectly
signed.

Still todo - finish up building the MatchingDatasetAssertion, build an
AuthnStatementAssertion, and finish building / signing the Response.

Authors: @richardtowers @sfkamath