5225: Build, Sign and Encrypt Assertions and Sign Response
Fixed the building of the MatchingDatasetAssertion (needed to use some
code from saml-utils due to some wierd issue marshalling Addresses built
using the opensaml APIs).

Built the AuthnStatementAssertion containing LevelOfAssurance and the IP
address.

We've added a new keystore for signing Responses we send to Verify in
addition to the existing keystore which we use for signing AuthnRequests
sent to eIDAS. It makes sense for these to be different keystores
because we'll need to use a cert that chains with Verify's PKI when
talking to Verify.

Extracted a SigningHelper to sign arbitrary XMLObjects to avoid
duplicating that code.

TODO - at the moment the address we're putting in the MDS Assertion
isn't really right. This is due to the eIDAS stub IdP not sending us an
address in the expected format.

TODO - We're not using the ResponseFormGenerator in our Resource at the
moment - consider either using it or removing it if it isn't needed.

Authors: @richardtowers @sfkamath