5225: Configure jboss.logging to use slf4j
jboss.logging is a dependency of the hibernate validator, which is a
dependency of dropwizard validation. Dropwizard uses slf4j as its main
logging provider, but jboss.logging defaults to log4j. This causes a
warning to be printed to the console on application startup that:

> log4j:WARN No appenders could be found for logger (org.jboss.logging)

It makes sense to configure jboss.logging to use the same logger as
dropwizard, rather than configuring it independently. This can be done
with a system property as described here:

https://github.com/spring-projects/spring-boot/issues/1928#issuecomment-63131908

This can be set using a static block, as described the first update of this
answer: http://stackoverflow.com/a/19488546/1344760

Authors: @richardtowers @sfkamath