bau: Update the stub-idp-cert
Also update the .env file so that it's still possible to run the bridge
locally. This required the addition of a script to fish into the
manifest files (assuming they're in a directory next to this one) and
pluck out the one keystore we don't want to open source (because it
contains keys from the verify PKI).

Authors: @richardtowers