5413: Remove crown and add alpha banner
This service is not yet live. To prevent people stumbling across it and
thinking it's part of .gov.uk because of the branding we've removed the
crown and added a feedback link.

For now the feedback link is just a mailto to eidas-alpha, this wouldn't
be good enough for a real service, but it's better than nothing.

Authors: @richardtowers @sfkamath