bau: use country picker to select destination node
This change introduces a CountryRepository which provides a map of
country code to entity id.

This is used after selecting a country to find the correct entity ID
that an AuthnRequest should be sent to. In addition, the retrieved
entity ID is stored in the session token so that it can be retrieved
when we receive a response later.

After receiving the Response the entity ID in the token is now used to
verify it by checking the Issuer against the value.

This means that `EidasAuthnRequestGenerator` and `ResponseHandler` no
longer require a `eidasEntityId` constructor parameter so the related
config has been removed.