EID-1: Skip signature validation on broken role descriptors
There's a bug with the eIDAS reference node (EID-108) where it
incorrectly signs RoleDescriptors in its metadata. This bug is present
in some of the other member states' nodes, so we need to work around it.

This commit adds a config flag to eidasbridge.yml for a country
"workaroundBrokenRoleDescriptorSignatures", if set, this will use a
custom SignatureValidationFilter which is aware of the bug and will skip
signature validation on RoleDescriptors which have empty SignatureValues
(which is the issue).

We don't think that adding this behaviour is a security problem because:

1) The EntityDescriptor still needs to be signed correctly, so it's
   still not possible to edit the Metadata.
2) The eIDAS reference implementation doesn't check these signatures at
   all.
3) We're only disabling validation for countries that have the bug.
4) We're only skipping RoleDescriptors that have signatures broken in
   the manner caused by the bug (so other broken signatures would still
   be invalid).
5) Verify doesn't sign RoleDescriptors in its metadata at all.

We'll need to check the member states that this is enabled for
periodically and remove the workaround once their code is fixed.

Authors: @richardtowers @sfkamath