bau: don't attempt to set environment variables unless run is invoked
On jenkins we don't have the verify-eidas-bridge-manifests folder
checked out, so ./gradlew run won't work. That's fine, because jenkins
only needs to run the tests. Unfortunately, gradle will still try and
read the file for the run task, even if it's not being invoked. To avoid
this we've added a conditionaly so we only try and set the environment
up for the run task if we're actually running it.

Authors: @richardtowers @sfkamath