bau: use MetadataCredentialResolver to pick enc. key
This uses the opensaml 'way' to retrieve the encrpytion public key from
metadata. In order to minimise changes to our saml libs I have kept have
not changed any other classes interfaces which means that our Retriever
object will retrieve a Credential then extract the public key and return it
where it will be wrapped in a new Credential by the calling object.