EID-94: Ensure our metadata is valid against the md schema
Austria pointed out that our AssertionConsumerService was missing an
index, and was therefore not valid against the metadata schema. This
commit fixes the issue by adding an index of 0 and adds a test to ensure
the metadata is valid.

I've downloaded the schema files and changed their imports to point at
local files, rather than URLs, so the test is not dependent on an
internet connection.

Authors: @richardtowers @sfkamath